Reorganize into multi-project
It all lives inside spark-docker for the moment, could be moved later,
we'll see.
