+++ spark-docker/common/src/main/scala/types/StructureTypes.scala
